fix: Address review feedback — cleanup stale fork_bound_threads on name reuse
- Add fork_bound_threads.remove() in cleanup_coworker_state to prevent
  stale thread bindings from bleeding into reused coworker names
- Update docs/architecture.md: fork_bound_threads is now rebuilt on
  startup (not ephemeral for task-spawned coworkers)
- Update fork_bound_threads doc comment with all three population paths
  and cleanup path
- Replace bare HashMap tests with DaemonPersistentState-based tests
  that exercise the actual storage path
- Move rpc_task tests to separate file per CLAUDE.md convention
