+++ src/daemon/rpc_task.rs
@@ -584,236 +584,6 @@ pub(super) fn handle_task_claim(
 // Tests
 // ============================================================================
 
+#[path = "rpc_task_tests.rs"]
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_apply_task_channel_mapping_sets_channel() {
-        let mut map = HashMap::new();
-        let changed = apply_task_channel_mapping(&mut map, "42", Some("auth"), false);
-        assert!(changed);
-        assert_eq!(map.get("42"), Some(&"auth".to_string()));
-    }
-
-    #[test]
-    fn test_apply_task_channel_mapping_overwrites_existing() {
-        let mut map = HashMap::new();
-        map.insert("42".to_string(), "old-channel".to_string());
-        let changed = apply_task_channel_mapping(&mut map, "42", Some("new-channel"), false);
-        assert!(changed);
-        assert_eq!(map.get("42"), Some(&"new-channel".to_string()));
-    }
-
-    #[test]
-    fn test_apply_task_channel_mapping_ignores_none() {
-        let mut map = HashMap::new();
-        map.insert("42".to_string(), "auth".to_string());
-        let changed = apply_task_channel_mapping(&mut map, "42", None, false);
-        assert!(!changed);
-        assert_eq!(map.get("42"), Some(&"auth".to_string()));
-    }
-
-    #[test]
-    fn test_apply_task_channel_mapping_ignores_empty_without_clear() {
-        let mut map = HashMap::new();
-        map.insert("42".to_string(), "auth".to_string());
-        // On create (allow_clear=false), empty string is ignored
-        let changed = apply_task_channel_mapping(&mut map, "42", Some(""), false);
-        assert!(!changed);
-        assert_eq!(map.get("42"), Some(&"auth".to_string()));
-    }
-
-    #[test]
-    fn test_apply_task_channel_mapping_clears_with_empty_on_update() {
-        let mut map = HashMap::new();
-        map.insert("42".to_string(), "auth".to_string());
-        // On update (allow_clear=true), empty string clears the mapping
-        let changed = apply_task_channel_mapping(&mut map, "42", Some(""), true);
-        assert!(changed);
-        assert!(!map.contains_key("42"));
-    }
-
-    #[test]
-    fn test_apply_task_channel_mapping_clear_nonexistent_is_noop() {
-        let mut map = HashMap::new();
-        // Clearing a mapping that doesn't exist returns false (no state modification)
-        let changed = apply_task_channel_mapping(&mut map, "99", Some(""), true);
-        assert!(!changed);
-        assert!(map.is_empty());
-    }
-
-    #[test]
-    fn test_apply_task_channel_mapping_none_on_empty_map() {
-        let mut map: HashMap<String, String> = HashMap::new();
-        let changed = apply_task_channel_mapping(&mut map, "42", None, true);
-        assert!(!changed);
-        assert!(map.is_empty());
-    }
-
-    #[test]
-    fn test_validate_model_format_valid() {
-        assert!(validate_model_format("claude/opus").is_ok());
-        assert!(validate_model_format("claude/sonnet").is_ok());
-        assert!(validate_model_format("claude/haiku").is_ok());
-        assert!(validate_model_format("codex/o3").is_ok());
-        assert!(validate_model_format("codex/o4-mini").is_ok());
-    }
-
-    #[test]
-    fn test_validate_model_format_invalid() {
-        // Missing slash
-        assert!(validate_model_format("claude-opus").is_err());
-        // Multiple slashes
-        assert!(validate_model_format("claude/opus/extra").is_err());
-        // Empty string
-        assert!(validate_model_format("").is_err());
-        // Only slash
-        assert!(validate_model_format("/").is_err());
-        // Empty provider
-        assert!(validate_model_format("/opus").is_err());
-        // Empty model
-        assert!(validate_model_format("claude/").is_err());
-        // Unsupported provider
-        assert!(validate_model_format("unknown/opus").is_err());
-        assert!(validate_model_format("openai/gpt4").is_err());
-        // Whitespace in model or provider
-        assert!(validate_model_format("claude/ opus").is_err());
-        assert!(validate_model_format("claude /opus").is_err());
-        assert!(validate_model_format(" claude/opus").is_err());
-        assert!(validate_model_format("claude/opus ").is_err());
-    }
-
-    #[test]
-    fn test_apply_task_model_mapping_sets_model() {
-        let mut map = HashMap::new();
-        let changed = apply_task_model_mapping(&mut map, "42", Some("claude/opus"), false);
-        assert!(changed.is_ok());
-        assert!(changed.unwrap());
-        assert_eq!(map.get("42"), Some(&"claude/opus".to_string()));
-    }
-
-    #[test]
-    fn test_apply_task_model_mapping_rejects_invalid_format() {
-        let mut map = HashMap::new();
-        let result = apply_task_model_mapping(&mut map, "42", Some("invalid-format"), false);
-        assert!(result.is_err());
-        assert!(map.is_empty());
-    }
-
-    #[test]
-    fn test_apply_task_model_mapping_overwrites_existing() {
-        let mut map = HashMap::new();
-        map.insert("42".to_string(), "claude/opus".to_string());
-        let changed =
-            apply_task_model_mapping(&mut map, "42", Some("claude/sonnet"), false).unwrap();
-        assert!(changed);
-        assert_eq!(map.get("42"), Some(&"claude/sonnet".to_string()));
-    }
-
-    #[test]
-    fn test_apply_task_model_mapping_ignores_none() {
-        let mut map = HashMap::new();
-        map.insert("42".to_string(), "claude/opus".to_string());
-        let changed = apply_task_model_mapping(&mut map, "42", None, false).unwrap();
-        assert!(!changed);
-        assert_eq!(map.get("42"), Some(&"claude/opus".to_string()));
-    }
-
-    #[test]
-    fn test_apply_task_model_mapping_ignores_empty_without_clear() {
-        let mut map = HashMap::new();
-        map.insert("42".to_string(), "claude/opus".to_string());
-        // On create (allow_clear=false), empty string is ignored
-        let changed = apply_task_model_mapping(&mut map, "42", Some(""), false).unwrap();
-        assert!(!changed);
-        assert_eq!(map.get("42"), Some(&"claude/opus".to_string()));
-    }
-
-    #[test]
-    fn test_apply_task_model_mapping_clears_with_empty_on_update() {
-        let mut map = HashMap::new();
-        map.insert("42".to_string(), "claude/opus".to_string());
-        // On update (allow_clear=true), empty string clears the mapping
-        let changed = apply_task_model_mapping(&mut map, "42", Some(""), true).unwrap();
-        assert!(changed);
-        assert!(!map.contains_key("42"));
-    }
-
-    #[test]
-    fn test_apply_task_model_mapping_clear_nonexistent_is_noop() {
-        let mut map = HashMap::new();
-        // Clearing a mapping that doesn't exist returns false (no state modification)
-        let changed = apply_task_model_mapping(&mut map, "99", Some(""), true).unwrap();
-        assert!(!changed);
-        assert!(map.is_empty());
-    }
-
-    #[test]
-    fn test_apply_task_model_mapping_none_on_empty_map() {
-        let mut map: HashMap<String, String> = HashMap::new();
-        let changed = apply_task_model_mapping(&mut map, "42", None, true).unwrap();
-        assert!(!changed);
-        assert!(map.is_empty());
-    }
-
-    // ── task_thread_id tests ──────────────────────────────────────────────────
-
-    /// When `thread_id` is provided, it is stored in `task_thread_id`.
-    #[test]
-    fn test_task_thread_id_stored_when_provided() {
-        let mut task_thread_id: HashMap<String, String> = HashMap::new();
-        let thread_id = Some("thread-parent-uuid-abc");
-        if let Some(tid) = thread_id {
-            task_thread_id.insert("42".to_string(), tid.to_string());
-        }
-        assert_eq!(
-            task_thread_id.get("42"),
-            Some(&"thread-parent-uuid-abc".to_string())
-        );
-    }
-
-    /// When `thread_id` is `None`, nothing is stored in `task_thread_id`.
-    #[test]
-    fn test_task_thread_id_not_stored_when_none() {
-        let mut task_thread_id: HashMap<String, String> = HashMap::new();
-        let thread_id: Option<&str> = None;
-        if let Some(tid) = thread_id {
-            task_thread_id.insert("42".to_string(), tid.to_string());
-        }
-        assert!(task_thread_id.is_empty());
-    }
-
-    /// `DaemonPersistentState` with `task_thread_id` survives round-trip JSON
-    /// serialization (backward-compatible with states that lack the field).
-    #[test]
-    fn test_task_thread_id_serialization_roundtrip() {
-        use crate::daemon::state::DaemonPersistentState;
-
-        let mut ps = DaemonPersistentState::default();
-        ps.task_thread_id
-            .insert("99".to_string(), "thread-parent-xyz".to_string());
-
-        let json = serde_json::to_string(&ps).expect("serialize");
-        let parsed: DaemonPersistentState = serde_json::from_str(&json).expect("deserialize");
-        assert_eq!(
-            parsed.task_thread_id.get("99"),
-            Some(&"thread-parent-xyz".to_string())
-        );
-    }
-
-    /// Older daemon state JSON (without `task_thread_id`) deserializes with an
-    /// empty map (the `#[serde(default)]` attribute).
-    #[test]
-    fn test_task_thread_id_defaults_to_empty_on_old_state() {
-        use crate::daemon::state::DaemonPersistentState;
-
-        // JSON payload that represents a state without task_thread_id
-        let json = r#"{}"#;
-        let ps: DaemonPersistentState = serde_json::from_str(json).expect("deserialize");
-        assert!(
-            ps.task_thread_id.is_empty(),
-            "task_thread_id should default to empty for old state files"
-        );
-    }
-}
+mod tests;
